pies/pi1/pyproject.toml
[project]
name = "pi1"
version = "2026.1.0"
description = "Add your description here"
requires-python = ">=3.10"
dependencies = [
    "gpiohero",
    "paho-mqtt>=2.1.0",
]

[tool.uv.sources]
gpiohero = { workspace = true }
